07.10  커밋 목록 / 작성자 : 성근 ---------------------------------------- * 추가 - 인증글 메뉴 팝업 - 챌린지 인증글 삭제 기능 - Board 객체에 masterToken 추가(인증글 승인/반려 용) ---------------------------------------- * 삭제 - ---------------------------------------- * 수정 - (피드, 댓글, 인증글) 삭제 시 확인창 출력되게

+++ app/src/main/java/com/example/bookworm/bottomMenu/Feed/CustomPopup.kt
@@ -1,6 +1,8 @@
 package com.example.bookworm.Feed
 
+import android.app.AlertDialog
 import android.content.Context
+import android.content.DialogInterface
 import android.content.Intent
 import android.view.MenuItem
 import android.view.View
@@ -27,7 +29,7 @@ class CustomPopup(context: Context?, anchor: View?) : PopupMenu(context, anchor)
     var commentitem: Comment? = null
     var boardFB: BoardFB? = null
 
-    fun setItems(context: Context, fbModule: FBModule, feed: Feed) {
+    fun setItems(context: Context, fbModule: FBModule, feed: Feed) { //피드탭
         this.fbModule = fbModule
         this.context2 = context
         this.item = feed
@@ -35,7 +37,15 @@ class CustomPopup(context: Context?, anchor: View?) : PopupMenu(context, anchor)
         this.menuInflater.inflate(R.menu.feed_menu, this.menu) //레이아웃에 inflate
     }
 
-    fun setItems(context: Context, fbModule: FBModule, comment: Comment, feed: Feed) {
+    fun setItems(context: Context, boardFB: BoardFB, board: Board) { //인증글 댓글화면 (인증글)
+        this.boardFB = boardFB
+        this.context2 = context
+        this.boardItem = board
+        this.layout = R.menu.board_menu
+        this.menuInflater.inflate(R.menu.board_menu, this.menu) //레이아웃에 inflate
+    }
+
+    fun setItems(context: Context, fbModule: FBModule, comment: Comment, feed: Feed) { //피드 댓글화면
         this.fbModule = fbModule
         this.context2 = context
         this.item = feed
@@ -44,7 +54,12 @@ class CustomPopup(context: Context?, anchor: View?) : PopupMenu(context, anchor)
         this.menuInflater.inflate(R.menu.comment_menu, this.menu) //레이아웃에 inflate
     }
 
-    fun setItems(context: Context, boardFB: BoardFB, comment: Comment, board: Board) {
+    fun setItems(
+        context: Context,
+        boardFB: BoardFB,
+        comment: Comment,
+        board: Board
+    ) { //인증글 댓글화면 (댓글)
         this.boardFB = boardFB
         this.context2 = context
         this.boardItem = board
@@ -79,11 +94,18 @@ class CustomPopup(context: Context?, anchor: View?) : PopupMenu(context, anchor)
                     return true
                 }
                 R.id.menu_delete -> {
-                    fbModule!!.deleteData(1, item!!.feedID) //삭제
-                    oldList?.removeAt(pos)
-                    ff.feedAdapter!!.submitList(oldList)
-                    //만약 댓글을 모아보는 액티비티(subactivity_comment)에 있는 경우, 해당 액티비티를 종료
-                    if (context is subactivity_comment) (context as subactivity_comment).finish()
+                    AlertDialog.Builder(context)
+                        .setMessage("정말 삭제하시겠습니까?")
+                        .setPositiveButton("네") { dialog: DialogInterface, which: Int ->
+                            dialog.dismiss()
+                            fbModule!!.deleteData(1, item!!.feedID) //삭제
+                            oldList?.removeAt(pos)
+                            ff.feedAdapter!!.submitList(oldList)
+                            //만약 댓글을 모아보는 액티비티(subactivity_comment)에 있는 경우, 해당 액티비티를 종료
+                            if (context is subactivity_comment) (context as subactivity_comment).finish()
+                        }
+                        .setNegativeButton("아니요") { dialog: DialogInterface, which: Int -> dialog.dismiss() }
+                        .show()
                 }
                 else -> return true
             }
@@ -93,14 +115,21 @@ class CustomPopup(context: Context?, anchor: View?) : PopupMenu(context, anchor)
             var oldList: ArrayList<Any> = ArrayList(ac!!.commentList)
             when (p0?.itemId) {
                 R.id.menu_delete -> {
-                    fbModule!!.deleteData(1, item!!.feedID, commentitem!!.commentID) //삭제
-                    val data = HashMap<String, Comment>()
-                    data.put("comment", commentitem!!)
-                    CommentsCounter()
-                        .removeCounter(data, context, item!!.feedID)
+                    AlertDialog.Builder(context)
+                        .setMessage("정말 삭제하시겠습니까?")
+                        .setPositiveButton("네") { dialog: DialogInterface, which: Int ->
+                            dialog.dismiss()
+                            fbModule!!.deleteData(1, item!!.feedID, commentitem!!.commentID) //삭제
+                            val data = HashMap<String, Comment>()
+                            data.put("comment", commentitem!!)
+                            CommentsCounter()
+                                .removeCounter(data, context, item!!.feedID)
 
-                    oldList?.removeAt(pos)
-                    (context as (subactivity_comment)).replaceItem(oldList)
+                            oldList?.removeAt(pos)
+                            (context as (subactivity_comment)).replaceItem(oldList)
+                        }
+                        .setNegativeButton("아니요") { dialog: DialogInterface, which: Int -> dialog.dismiss() }
+                        .show()
                 }
                 else -> return true
             }
@@ -111,19 +140,47 @@ class CustomPopup(context: Context?, anchor: View?) : PopupMenu(context, anchor)
             var oldList: ArrayList<Any> = ArrayList(ac!!.commentList)
             when (p0?.itemId) {
                 R.id.menu_delete -> {
-                    boardFB!!.deleteComment(boardItem, commentitem!!.commentID) //삭제
-                    val data = HashMap<String, Comment>()
-                    data.put("comment", commentitem!!)
-                    Board_CommentsCounter()
-                        .removeCounter(
-                            data,
-                            context,
-                            boardItem!!.challengeName,
-                            boardItem!!.boardID
-                        )
+                    AlertDialog.Builder(context)
+                        .setMessage("정말 삭제하시겠습니까?")
+                        .setPositiveButton("네") { dialog: DialogInterface, which: Int ->
+                            dialog.dismiss()
+                            boardFB!!.deleteComment(boardItem, commentitem!!.commentID) //삭제
+                            val data = HashMap<String, Comment>()
+                            data.put("comment", commentitem!!)
+                            Board_CommentsCounter()
+                                .removeCounter(
+                                    data,
+                                    context,
+                                    boardItem!!.challengeName,
+                                    boardItem!!.boardID
+                                )
 
-                    oldList?.removeAt(pos)
-                    (context as (subactivity_challenge_board_comment)).replaceItem(oldList)
+                            oldList?.removeAt(pos)
+                            (context as (subactivity_challenge_board_comment)).replaceItem(oldList)
+                        }
+                        .setNegativeButton("아니요") { dialog: DialogInterface, which: Int -> dialog.dismiss() }
+                        .show()
+                }
+                else -> return true
+            }
+        } else if (layout == R.menu.board_menu) { //인증글의 메뉴 팝업
+            when (p0?.itemId) {
+                R.id.menu_allow -> { //인증글 승인시 (챌린지 개설자 권한)
+
+                }
+                R.id.menu_notallow -> { //인증글 반려시 (챌린지 개설자 권한)
+
+                }
+                R.id.menu_delete -> { //본인이 인증글 삭제시
+                    AlertDialog.Builder(context)
+                        .setMessage("정말 삭제하시겠습니까?")
+                        .setPositiveButton("네") { dialog: DialogInterface, which: Int ->
+                            dialog.dismiss()
+                            boardFB!!.deleteBoard(boardItem) //인증글 삭제
+                            (context as subactivity_challenge_board_comment).finish() //액티비티 종료
+                        }
+                        .setNegativeButton("아니요") { dialog: DialogInterface, which: Int -> dialog.dismiss() }
+                        .show()
                 }
                 else -> return true
             }
@@ -135,9 +192,21 @@ class CustomPopup(context: Context?, anchor: View?) : PopupMenu(context, anchor)
         if (layout == R.menu.feed_menu) { //레이아웃이 피드_메뉴 일때
             this.menu.findItem(R.id.menu_delete).setVisible(boolean)
             this.menu.findItem(R.id.menu_modify).setVisible(boolean)
-        } else if (layout == R.menu.comment_menu) { //레이아웃이 댓글_메뉴 일때
+        }
+        else if (layout == R.menu.comment_menu) { //레이아웃이 댓글_메뉴 일때
             this.menu.findItem(R.id.menu_delete).setVisible(boolean)
-        } else if (layout == R.menu.board_comment_menu) { //레이아웃이 인증글_댓글_메뉴 일때
+        }
+        else if (layout == R.menu.board_menu) { //레이아웃이 인증글_메뉴 일때
+            this.menu.findItem(R.id.menu_allow).setVisible(boolean)
+            this.menu.findItem(R.id.menu_notallow).setVisible(boolean)
+        }
+        else if (layout == R.menu.board_comment_menu) { //레이아웃이 인증글_댓글_메뉴 일때
+            this.menu.findItem(R.id.menu_delete).setVisible(boolean)
+        }
+    }
+
+    fun  setDeleteVisible(boolean: Boolean) { //작성자만 자신의 인증글을 삭제할 수 있게 함
+        if (layout == R.menu.board_menu) { //레이아웃이 인증글_메뉴 일때
             this.menu.findItem(R.id.menu_delete).setVisible(boolean)
         }
     }
